Some more refactoring
