+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>de.teragam</groupId>
     <artifactId>jfxshader</artifactId>
-    <version>0.8.5</version>
+    <version>1.0.0</version>
 
     <properties>
         <maven.compiler.source>11</maven.compiler.source>
