--- src/main/java/de/teragam/jfxshader/material/IMaterialRenderer.java
@@ -1,4 +1,0 @@
-package de.teragam.jfxshader.material;
-
-public interface IMaterialRenderer {
-}
