--- src/main/java/de/teragam/jfxshader/material/MaterialRenderer.java
@@ -1,16 +1,0 @@
-package de.teragam.jfxshader.material;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Inherited;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Inherited
-@Documented
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface MaterialRenderer {
-    Class<? extends IMaterialRenderer> value();
-}
