+++ src/main/java/de/teragam/jfxshader/material/internal/es2/ES2ShaderMaterialPeerRenderer.java
@@ -1,10 +1,41 @@
 package de.teragam.jfxshader.material.internal.es2;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+import javafx.scene.image.Image;
+
 import com.sun.javafx.geom.transform.GeneralTransform3D;
+import com.sun.prism.Graphics;
+import com.sun.prism.Texture;
+import com.sun.prism.impl.ps.BaseShaderContext;
 
 import de.teragam.jfxshader.material.internal.AbstractShaderMaterialPeerRenderer;
+import de.teragam.jfxshader.material.internal.ShaderMeshView;
+import de.teragam.jfxshader.util.Reflect;
 
 public class ES2ShaderMaterialPeerRenderer extends AbstractShaderMaterialPeerRenderer {
+
+    @Override
+    public void render(Graphics g, ShaderMeshView meshView, BaseShaderContext context, Map<Integer, Image> imageIndexMap) {
+        if (!(meshView instanceof ES2ShaderMeshView)) {
+            throw new IllegalArgumentException("meshView must be an instance of ES2ShaderMeshView");
+        }
+        final List<Texture> textures = new ArrayList<>();
+        imageIndexMap.forEach((index, image) -> {
+            final Texture texture = this.getTexture(context, image, false);
+            Reflect.on(BaseShaderContext.class).method("updateTexture").invoke(context, index, texture);
+            if (texture != null) {
+                textures.add(texture);
+            }
+        });
+        final Reflect<? extends BaseShaderContext> contextReflect = Reflect.on(context.getClass());
+        final Object glContext = contextReflect.getFieldValue("glContext", context);
+        Reflect.on("com.sun.prism.es2.GLContext").method("renderMeshView").invoke(glContext, ((ES2ShaderMeshView) meshView).getNativeHandle());
+
+        textures.forEach(Texture::unlock);
+    }
 
     @Override
     protected float[] convertMatrix(GeneralTransform3D src) {
